Added request API

+++ demo/src/main/java/com/proj/tech/mapper/blocks/InstructionMapper.java
@@ -10,7 +10,7 @@ public class InstructionMapper {
                 instruction.getName(),
                 instruction.getInstruction(),
                 instruction.getCode().getId(),
-                instruction.getOrderInCode()
+                instruction.getOrderInCode().longValue()
         );
     }
 }
